+++ Observer/Customer/AfterAddressSave.php
@@ -22,11 +22,12 @@ class AfterAddressSave extends \Drip\Connect\Observer\Base
     public function __construct(
         \Drip\Connect\Helper\Data $connectHelper,
         \Magento\Framework\Registry $registry,
+        \Drip\Connect\Logger\Logger $logger,
         \Drip\Connect\Helper\Customer $customerHelper,
         \Magento\Framework\Serialize\Serializer\Json $json,
         \Magento\Customer\Model\CustomerFactory $customerCustomerFactory
     ) {
-        parent::__construct($connectHelper);
+        parent::__construct($connectHelper, $logger);
         $this->registry = $registry;
         $this->customerHelper = $customerHelper;
         $this->customerCustomerFactory = $customerCustomerFactory;